user-delete/requirements.txt
Flask==3.0.3
flasgger
pymongo
python-dotenv